rename user package to identity package
